+++ packages/adapters/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamba/adapters
 
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamba/core@0.16.0
+
 ## 0.15.0
 
 ### Patch Changes